Streamline trainer + new copy model inference / dataset / python Volume bindings  (#960)
* initial messy port

* huge cleanup

* more cleanup

* unused configs/conditionals

* cleaner

* and more ...

* c

* c2

* c3

* c4

* c5

* move target/batch prep to a class outside of trainer

* move some losses out of train file

* move rest of defaults to the config file

* viz helper / dedup

* consolidate checkpoint handling

* more config cleanup

* remove unused eager eval

* Update train_rowcol_cond.py

* move the config print into the config file

* Update .gitignore

* mmap tifxyz, dont load labels unless req

* Update profile_rowcol_cond_dataset.py

* Update dataset_rowcol_cond.py

* dedup/optim

* Update direction_helpers.py

* sparse normalize

* some more perf for the stream/flow model

* perf2

* Update dataset_rowcol_cond.py

* redo copy-out / speed up patch cache

* updates to patch finding / dataset

* Update chunk_finding.py

* streamline inference

* tifxyz output

* tqdm prog

* speed

* cleanup tmp files after writing output

* remove cpu sampling

* add iterations

* Update infer_streamline.py

* tta / vector geomedian

* fix meta

* bring displacement copy into updated dataset/trainer

* Update profile_rowcol_cond_dataset.py

* fix dir aug

* sampled displ loss

* Add distributed streamline inference

* infer_streamline: place new tifxyz pixels at each row's actual edge

`_build_merged_streamline_tifxyz` previously concatenated the integrated
strip globally onto the end (or start) of the base grid, so every extension
landed at a single global col/row even though each row's `E_R` (rightmost
valid col for cond="left" / etc.) varied. Whenever the base mesh's edge
was jagged — i.e. col `base.shape[1]-1` was invalid for the row — the
strip pixel had no 4-grid-neighbour to the base, producing tiny floating
islands disconnected from the main surface in the tifxyz output.

Verified on PHercParis4/w00_0805261021 (8x H100, BS=4, TTA+compile):
  - before: 22 connected components after iter 1 (main + 21 floaters)
  - after:  1 connected component, 1.0 largest_fraction
  - extension xyz is unchanged (same +440 valid points, same
    integration_active_endpoints=646, same low_validity=4106) — only
    placement in the grid is corrected.

Also:
  - MERGE_OUTPUTS_MMAP_DIR moved off /tmp (which is on root fs) onto
    /ephemeral by default; overridable via $VESUVIUS_MERGE_MMAP_DIR. The
    merger's tempfile.TemporaryDirectory leaks ~120-230 GB per crashed run
    (mmap-backed chunks); these were filling root and inducing SIGBUS on
    subsequent runs.
  - Per-iter `output_components` (4-conn count / top sizes /
    largest_fraction) added to the iteration summary so the
    floater regression can be detected end-to-end.


* infer_streamline: rectangular-crop preprocessing for jagged inputs

The streamline iterator's growth is dominated by the model's confidence at
the conditioning edge. When the input mesh has a jagged boundary on the
grow-direction side (e.g. PHercParis4/w00 has ER_std=68 cols on its right
boundary, with only 7% of rows reaching the global rightmost col), the
model becomes very conservative and saturates after 1-3 iters.

Preprocess: crop the input to the largest rectangular sub-region with a
uniform conditioning edge. The crop is direction-aware (uses
_get_growth_context to pick the conditioning side) and chosen to maximise
area (kept_lines * edge_extent). Floaters inside the rectangle are dropped
so the cropped mesh is single-component.

Empirically on PHercParis4/w00 grow-right (8x H100, BS=4, TTA+compile):
  - Jagged input (1587x446, ER std=68):
      iter1 +440 valid, iter2 +69, iter3 +8, then saturated. Total +517.
  - Auto-cropped input (1443x404, ER std=0):
      iter1 +916 (active_ep=1373), iter2 +699 (active_ep=1048),
      iter3 +next (active_ep=789) — sustained productive growth.

Default ON; toggle with --no-rectangular-crop. The crop is run only on
the first iter's input; subsequent iters use the natural extension output
(which by construction stays near-rectangular thanks to the
per-row strip placement landed in the previous commit).


* infer_streamline: thread-pool the merger's mean-finalize loop

The merger's per-chunk mean-finalize (read mmap sum_chunk + count_chunk,
np.divide, write zarr) is the dominant cost in the post-bbox-inference phase
on rank 0 — ~2 min/field over ~570 chunks for the velocity_dir /
surface_attract / trace_validity outputs. The chunks are independent, and
both the numpy and zarr-blosc operations release the GIL during their
heavy work, so a ThreadPoolExecutor with 16 workers cuts the per-field
finalize from ~2 min to ~10-20 s.

Worker count is configurable via VESUVIUS_FINALIZE_WORKERS (default 16).
No behaviour change for the geomedian path (also independent across
chunks but currently sequential — left untouched as we default to mean).

Smoke-tested end-to-end on the iter7 output of an ongoing infinite
extension (PHercParis4/w00, grow=right, BS=8, --no-tta, --compile):
12:50 iter wallclock (vs 17 min/iter on the same input with the same
extension semantics, single component, +1671 valid points).


* infer_streamline: expose --trace-validity-threshold CLI flag

The integrator rejects every candidate whose model-predicted
trace_validity (sigmoid) falls below TRACE_VALIDITY_THRESHOLD = 0.5.
Empirically, this is the dominant saturation reason on iterative
extension: per-iter low_validity rejection grows monotonically
(1827 → 4061 over 7 iters of w00/right) while every other failure
diagnostic stays at 0. Once active_endpoints hits 0, the script
loops forever producing identical outputs.

Expose the threshold so we can probe whether the model has lower-
confidence-but-still-useful candidates near the saturation boundary.
Module default unchanged (0.5).


* infer_streamline: truncate input_name in output uuid to dodge 255-char limit

When iterative resume is used (reflatten-extend loop), each cycle's input
tifxyz dir name embeds its parent's full uuid → names grow ~40 chars/cycle.
After ~5 cycles we hit Linux's ext4 255-char filename ceiling on
meta.json / x.tif and the run crashes with `OSError: [Errno 36] File name
too long`. Keep the first 24 chars of input_name plus an 8-char sha1
digest so the uuid stays bounded.


* re-add copy inference

* wrap inference

* proper merge

* Update infer_rowcol_triplet_wraps.py

* add cmake python target and set PIC for it

* remove tmp test

* remove a few more tmp tests

* fix imports for new patch finding

* dont mutate non-mmappable tifxyz in place

---------

+++ vesuvius/src/vesuvius/neural_tracing/autoreg_mesh/losses.py
@@ -779,8 +779,14 @@ def _occupancy_metric(outputs: dict, batch: dict) -> Tensor:
         pred_grid = pred_grid.reshape(grid_shape[0], grid_shape[1], 3)
         target_grid = batch["target_grid_local"][batch_idx].detach().cpu().numpy()
         crop_shape = tuple(int(v) for v in volume.shape[-3:])
-        pred_vox = torch.from_numpy(voxelize_surface_grid(pred_grid.astype("float32"), crop_shape)).to(device=device)
-        target_vox = torch.from_numpy(voxelize_surface_grid(target_grid.astype("float32"), crop_shape)).to(device=device)
+        pred_vox = torch.from_numpy(voxelize_surface_grid(pred_grid.astype("float32"), crop_shape)).to(
+            device=device,
+            dtype=torch.float32,
+        )
+        target_vox = torch.from_numpy(voxelize_surface_grid(target_grid.astype("float32"), crop_shape)).to(
+            device=device,
+            dtype=torch.float32,
+        )
         losses.append(F.binary_cross_entropy(pred_vox.clamp(1e-6, 1.0 - 1e-6), target_vox))
     if not losses:
         return torch.zeros((), device=device)
